backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
django-jsonfield
numpy
pillow
opencv-python
torch
torchvision